@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace CodelyTv\Tests;
 
 use CodelyTv\Codelyber;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
 final class CodelyberTest extends TestCase
 {
-    /** @test */
-    public function itShouldSayHelloWhenGreeting(): void
+    #[Test] public function itShouldSayHelloWhenGreeting(): void
     {
         $codelyber = new Codelyber("Javi");
 
-        self::assertEquals("CodelyTV", $codelyber->greet());
+        $this->assertEquals("CodelyTV", $codelyber->greet());
     }
 }
